requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.1
Django==2.0
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
image==1.5.16
itsdangerous==0.24
Jinja2==2.9.6
lxml==4.1.0
MarkupSafe==1.0
ncclient==0.5.3
nodejs==0.1.1
npm==0.1.1
olefile==0.44
optional-django==0.1.0
paramiko==2.3.1
Pillow==4.3.0
pyang==1.7.3
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.1.2
pytz==2017.3
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
slackclient==1.1.3
matplotlib==2.2.2
pandas==0.22.0